Remove commit during Pick stage of commit apply
When a conflict is encountered during a cherry-pick of a template
commit, the commit is not made to the current repository. For the token
replacement process, the commit was expected for determination of target
files.

Instead, the cherry-pick skips the commit and the token replacement uses
a diff to determine the target files. The commit on the current
repository now serves the purpose that the Amend stage used to.